requirements.txt
numpy==1.19.2
pandas==1.1.2
streamlit==0.71.0
pycaret==2.2.2
Pillow==8.0.1
